CI: Move tests into their own suites
As our tests are mostly set to run as individual suites due to
them being disruptive, lets separate them into individual suites
where it makes sense.

Also use the new generic test target in the makefile to simplify
the makefile and reduce the number of targets
