+++ ci/infra/testrunner/tests/test_remove_master.py
