+++ ci/infra/testrunner/tests/test_upgrade_plan_all_fine.py
@@ -1,0 +1,17 @@
+import pytest
+
+from tests.utils import setup_kubernetes_version
+
+
+@pytest.mark.disruptive
+def test_upgrade_plan_all_fine(setup, skuba, kubectl, platform):
+    """
+    Starting from a up-to-date cluster, check what cluster/node plan report.
+    """
+
+    setup_kubernetes_version(platform, skuba, kubectl)
+    out = skuba.cluster_upgrade_plan()
+
+    assert out.find(
+        "Congratulations! You are already at the latest version available"
+    ) != -1
